jmp_checker.py
This is my final python script using Ghidra APIs to list all the information of the basic block, its instructions and the jumps, i.e which is the previous block that has jmp to current block and to which blocks the current block has jmps. It also checks if a block start with a NOP or not.